% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arruma_boletins.R
\name{arruma_boletins}
\alias{arruma_boletins}
\title{Title}
\usage{
arruma_boletins(dir_in, dir_out, modelo = 1)
}
\arguments{
\item{dir_in}{Diretório onde estão os boletins}

\item{dir_out}{Diretório de saída dos boletins arrumados}

\item{modelo}{Modelo de estrutura do boletim}
}
\description{
Title
}
