% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arruma_dados_rigeo.R
\name{arruma_dados_rigeo}
\alias{arruma_dados_rigeo}
\title{Recupera dados geoquímicos do Rigeo}
\usage{
arruma_dados_rigeo(
  folhas = "inputs/campo/folhas.shp",
  dir_in = "inputs/projetos/",
  dir_out = "outputs/"
)
}
\arguments{
\item{folhas}{Articulação das folhas. Dica: gerar grid das folhas no QGIS com
o plugin DSGTools - Grid Algorithms - Gerar Grid Sistematicamente}

\item{dir_in}{Diretório onde estão os arqivos zip de prontos do DGM}

\item{dir_out}{Diretório para gravação das planilhas de dados}
}
\description{
Prepara os dados disponibilizados pelo SGB-CPRM. Estes dados
estão em parte na área de download do GEOSGB e em parte no link dos projetos
no RIGEO e compactados na extenção zip. A função abre este arquivos e coloca
num diretório temporário e depois salva arquivos csv para cada classe e para
todas as classes. É preciso ter o cuidado de só colocar para rodar arquivos
semelhantes aos do modelo. Arquivos excel devem xlsx e as máscaras das
tabelas de atributos devem ser padronizadas.
}
\examples{
#arruma_dados_rigeo()
}
