% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifica_amostra.R
\name{classifica_amostra}
\alias{classifica_amostra}
\title{Title}
\usage{
classifica_amostra(file = "outputs/sc_tidy.csv")
}
\arguments{
\item{file}{Arquivo dos dados analiticos brutos}
}
\description{
Title
}
