% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifica_amostra_ca.R
\name{classifica_amostra_ca}
\alias{classifica_amostra_ca}
\title{Classificação dos dados geoquímicos para cartas de anomalias}
\usage{
classifica_amostra_ca(file = "outputs/sc_tidy.csv")
}
\arguments{
\item{file}{Arquivo dos dados analiticos brutos}
}
\description{
A ferramenta permite transformar os dados do Rigeo e classificar as amostra
pelo método Tucker Inner Fence (TIF) valores acima de 75\% são considerados
anômalos. Usar a mesma máscara do arquivo de saída da função
arruma_dados_rigeo do sedimento de corrente.
}
