% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifica_metodos.R
\name{classifica_metodos}
\alias{classifica_metodos}
\title{Métodos de classificação de dados geoquímicos}
\usage{
classifica_metodos(tipo_proc, mtd_class, lito_bacia, elem_val, base, nbc = 10)
}
\arguments{
\item{tipo_proc}{Tipo de processamento 1 = geral 2 = por unidades}

\item{mtd_class}{}

\item{elem_val}{Elementos válidos}

\item{base}{Base de dados preparada}
}
\description{
Métodos de classificação de dados geoquímicos
}
