% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutter.R
\name{cutter}
\alias{cutter}
\title{Corta pelas quebras calculadas do boxplot}
\usage{
cutter(x, cuts)
}
\arguments{
\item{x}{vetor}

\item{cuts}{quebra}
}
\description{
Corta pelas quebras calculadas do boxplot
}
