% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estrutura_diretorio.R
\name{estrutura_diretorio}
\alias{estrutura_diretorio}
\title{Cria diretórios e subdiretórios para o processamento de dados}
\usage{
estrutura_diretorio(path)
}
\arguments{
\item{path}{diretório do projeto}
}
\value{
Cria a estrutura dos diretórios onde devem ser lidos os arquivos de
entrada e salvos os arquivos de saída
}
\description{
Cria diretórios e subdiretórios para o processamento de dados
}
