% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gera_estacoes.R
\name{gera_estacoes}
\alias{gera_estacoes}
\title{Gera estações nas confluências dos rios}
\usage{
gera_estacoes(
  dem,
  dir_out = "outputs/",
  bases_model,
  EPSG = 4326,
  threshold = 250,
  min_length = "0.02",
  max_ordem = 4,
  wbt_wd = "outputs/modelo/"
)
}
\arguments{
\item{dem}{Modeolo digital do terreno (SRTM)}

\item{dir_out}{Diretório de saída}

\item{bases_model}{Bases definidas na função gera_bases_model}

\item{EPSG}{Sistema de coordenadas}

\item{threshold}{Valor de corte da área das bacias modeladas}

\item{min_length}{Comprimento mínimo do curso do rio}

\item{max_ordem}{Ordem Strhaler máxima para a validaçãon da estação}

\item{wbt_wd}{}
}
\value{
Imagem (.png) do mapa de planejamento preliminar.
Arquivo da shape (shp) das estações.
}
\description{
Desenvolvido para automatizar o processo de escolha do locais de amostragem
que devem ser a 100 m acima dos pontos de confluência dos rios. Também são
excluídos locais nas áreas urbanas e de massa d'água. Todos os arquivos devem
ser no formato shapefile.
}
\examples{
#gera_estacoes()

}
