% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legenda_boxplot.R
\name{legenda_boxplot}
\alias{legenda_boxplot}
\title{Legenda boxplot para o mapa TIF}
\usage{
legenda_boxplot(data = "mydata.csv", job_info = "myjob.csv", eq = "Ni")
}
\arguments{
\item{data}{Dados analíticos em csv}

\item{job_info}{Informações das condições analíticas}

\item{eq}{Elemento químico}
}
\description{
Legenda boxplot para o mapa TIF
}
