% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapa_tif_unidades.R
\name{mapa_tif_unidades}
\alias{mapa_tif_unidades}
\title{Mapa geoquímico (TIF)}
\usage{
mapa_tif_unidades(
  bacias,
  geologia,
  dados_transf_sc,
  info_bol,
  mylitho_max_sc,
  elem_val,
  leg2
)
}
\arguments{
\item{leg2}{}
}
\description{
Mapa geoquímico (TIF)
}
