% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modela_bacias.R
\name{modela_bacias}
\alias{modela_bacias}
\title{Extrai a drenagem do SRTM}
\usage{
modela_bacias(
  fase = 2,
  EPSG = 4326,
  dem,
  bases_model,
  gera_est,
  ex_campo,
  classe_am,
  bacia_minima = 4,
  bacia_maxima = 100,
  threshold = 250,
  snap_dist = "0.02",
  min_length = "0.02",
  max_ordem = 3,
  dir_out = "outputs/",
  wbt_wd = "outputs/modelo/"
)
}
\arguments{
\item{fase}{Fase do levantamento geoquímico (1= Planejamento, 2 = Pós campo)}

\item{EPSG}{crs da camada}

\item{dem}{Modelo digital de terreno (ex. SRTM)}

\item{bases_model}{Bases para a modelagem das bacias e drenagem pela
função gera_bases_model}

\item{gera_est}{}

\item{ex_campo}{Dados de campo extraídos pela função extrai_dados_campo}

\item{classe_am}{Classe da amostra: 1 = concentrado de bateia, 2 = sedimento de
corrente, 3 = rocha, 4 = solo}

\item{bacia_minima}{Área mínima das bacias planejadas}

\item{bacia_maxima}{Área máxima das bacias planejadas}

\item{threshold}{Numero de células mínimo para o traçado das drenagens}

\item{snap_dist}{Deslocamento máximo do ponto até a drenagem}

\item{min_length}{Comprimento mínimo da drenagem}

\item{max_ordem}{Máxima ordem do rio para a busca do snap point}

\item{dir_out}{Diretório de saída}
}
\value{
Rasters da preparação das imagens para a extração do modelo de
drenagem. Shapes e rasters das drenagens e trechos drenagems com as ordens
de strahler.
}
\description{
A partir da imagem SRTM, rios, massa d'água é gerado um modelo de drenagem.
A densidade de drenagem é dada pelo valor de threshold, quanto maior
este valor, menor a densidade dos cursos dos rios.
}
\examples{
# modelo_bacias()
}
