% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_dados_brutos.R
\name{prep_dados_brutos}
\alias{prep_dados_brutos}
\title{Prepara os dados brutos}
\usage{
prep_dados_brutos(data)
}
\arguments{
\item{data}{dados brutos}
}
\description{
Padroniza os LDs para o mínimo
}
