% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recorta_feicao_area.R
\name{recorta_feicao_area}
\alias{recorta_feicao_area}
\title{Recorta a feição pela área}
\usage{
recorta_feicao_area(
  dir_in = "inputs/campo/",
  dir_out = "outputs",
  area = "area_srtm",
  feicao_in = "geologia_br",
  feicao_out = "geologia_area"
)
}
\arguments{
\item{dir_in}{Diretório de entrada das feições}

\item{dir_out}{Diretório da feição de saída}

\item{area}{Nome do arquivo shape da área}

\item{feicao_in}{Nome da shape da feição de entrada}

\item{feicao_out}{Nome da shape da feição de saída}
}
\description{
Ferramenta que extrai a shape da feição (geologia, solo...) a partir de uma
feição de escala menor. Usei o mapa geológico do Brasil ao milhonésimo do SGB.
}
