% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testes_statisticos.R
\name{testes_estatisticos}
\alias{testes_estatisticos}
\title{Testes Estatísticos}
\usage{
testes_estatisticos(
  dados_transf_sc,
  dados_brutos_sc,
  info_bol,
  mylitho_max_sc,
  mtd_transf = 1
)
}
\arguments{
\item{info_bol}{Incormações das condições analíticas}

\item{mylitho_max_sc}{Litologia dominante de cada amostra considerando sua
bacia de captação}

\item{mydata}{Dados químicos transformados}

\item{mydata_b}{Dados quimícos brutos}
}
\description{
Testes de normalidade e de comparação das amostras
}
