% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforma_dados_05ld.R
\name{transforma_dados_05ld}
\alias{transforma_dados_05ld}
\title{Prepara e transforma dados para o tratamento estatístico}
\usage{
transforma_dados_05ld(data)
}
\arguments{
\item{data}{}
}
\description{
Transforma dados <LD para 0,5*LD
}
