% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valida_campo.R
\name{valida_campo}
\alias{valida_campo}
\title{Title}
\usage{
valida_campo(
  path_ctr,
  path_base,
  base,
  layer_base,
  amostras,
  threshold,
  buffer
)
}
\arguments{
\item{path_ctr}{diretório dos pontos de controle}

\item{path_base}{diretório da base de campo}

\item{base}{nome do gdb}

\item{layer_base}{feição das estações}

\item{amostras}{tabela com dados das amostras}

\item{threshold}{limiar em metros das distâcias entre estações}

\item{buffer}{raio de busca em relação ao ponto controle}
}
\value{
lista com as validações
}
\description{
Title
}
