% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validacao_elementos.R
\name{validacao_elementos}
\alias{validacao_elementos}
\title{Validação dos elementos químicos analisados}
\usage{
validacao_elementos(dir_out, base, corte_val)
}
\arguments{
\item{dir_out}{Diretório de saída}

\item{base}{Lista obtida da função prepara_bases}

\item{corte_val}{Critério de corte do percentual de dados censurados}
}
\description{
Validação dos elementos químicos analisados
}
