% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_teste_wilcoxon.R
\name{vis_teste_wilcox}
\alias{vis_teste_wilcox}
\title{Visual do Teste Wilcoxon}
\usage{
vis_teste_wilcox(dir_out, elem_val, base, lito_bacia, nbc)
}
\arguments{
\item{dir_out}{diretório de saída}

\item{elem_val}{elementos válidados}

\item{base}{bases de dados}

\item{lito_bacia}{bacias rotuladas}

\item{nbc}{número de bacias por unidade}
}
\description{
Visual do Teste Wilcoxon
}
