% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arruma_dados_rigeo.R
\name{arruma_dados_rigeo}
\alias{arruma_dados_rigeo}
\title{Recupera dados geoquímicos do Rigeo}
\usage{
arruma_dados_rigeo(
  folhas = "inputs/campo/folhas.shp",
  dir_in = "inputs/projetos/",
  dir_out = "outputs/"
)
}
\arguments{
\item{folhas}{Articulação das folhas. Dica: gerar grid das folhas no QGIS com
o plugin DSGTools - Grid Algorithms - Gerar Grid Sistematicamente}

\item{dir_in}{Diretório onde estão os arqivos zip de protos do DGM}

\item{dir_out}{Diretório para gravação das planilhas de dados}
}
\description{
Recupera dados geoquímicos do Rigeo
}
\examples{
#arruma_dados_rigeo()
}
