% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arruma_os.R
\name{arruma_os}
\alias{arruma_os}
\title{Arruma dados da OS}
\usage{
arruma_os(dir_os = "Data/PG2024/Rocha/os_solicitada/")
}
\arguments{
\item{dir_os}{Diretório dos documentos de solicitação das OS}
}
\description{
Arruma dados da OS
}
