% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifica_amostra.R
\name{classifica_amostra}
\alias{classifica_amostra}
\title{Classificação dos dados geoquímicos para cartas de anomalias}
\usage{
classifica_amostra(file = "outputs/sc_tidy.csv")
}
\arguments{
\item{file}{Arquivo dos dados analiticos brutos}
}
\description{
A ferramenta permite transformar os dados do Rigeo e classificar as amostra
pelo método Tucker Inner Fence (TIF) valores acima de 75\% são considerados
anômalos.
}
