% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifica_metodos.R
\name{classifica_metodos}
\alias{classifica_metodos}
\title{Métodos de classificação de dados geoquímicos}
\usage{
classifica_metodos(tipo_class, un_val, elem_val, base, list_break)
}
\arguments{
\item{tipo_class}{}
}
\description{
Métodos de classificação de dados geoquímicos
}
