% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digits.R
\name{digits}
\alias{digits}
\title{Calcula número de digitos decimais}
\usage{
digits(x, n = NULL, simplify = FALSE)
}
\arguments{
\item{x}{numero}

\item{simplify}{}
}
\description{
Calcula número de digitos decimais
}
