% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrai_dados_campo.R
\name{extrai_dados_campo}
\alias{extrai_dados_campo}
\title{Extrai dados de campo do banco de dados}
\usage{
extrai_dados_campo(tipo_base, dir_base, base_campo, dir_os)
}
\arguments{
\item{tipo_base}{Tipo de base de dados: 1 = FCAMPO, 2 = SURVEY123, 3 = QFIELD}

\item{dir_base}{Diretório da base de dados de campo}

\item{base_campo}{Nome da base de dados de campo}

\item{dir_os}{Diretório das planilhas de OS}
}
\value{
Retorna uma tabela com os dados de campo
}
\description{
Uma base de dados é preparada com as informações da localização das amostras
e os nomes de campo e laboratório. Esta função está implementada para
diferentes bases de dados: FCAMPO, SURVEY123 e QFIELD. Usa as planilhas de OS
(Ordem de Serviço) para recuperar a relação número de campo e número
de laboratório.
}
