% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gera_bases_model.R
\name{gera_bases_model}
\alias{gera_bases_model}
\title{Gera bases para a modelagem das bacias e estações}
\usage{
gera_bases_model(
  dir_in = "inputs/campo/",
  limite = "carta_100M.shp",
  limite_srtm = "area_srtm.shp",
  area_urbana = "area_urbana.shp",
  rios = "rios_ibge.shp",
  massa_dagua = "massa_dagua.shp"
)
}
\arguments{
\item{limite_srtm}{}
}
\description{
Converte as bases de entrada em um alista de dados para serem carregadas nas
funções modela_bacias e gera estações.
}
