% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gera_estacoes.R
\name{gera_estacoes}
\alias{gera_estacoes}
\title{Gera estações nas confluências dos rios}
\usage{
gera_estacoes(
  dir_shp = "inputs/campo/",
  dir_dem = "inputs/imagens/",
  dir_out = "outputs/",
  limite = "carta_100M.shp",
  area_urbana = "area_urbana.shp",
  estacoes = "estacoes_geradas.shp",
  srtm = "srtm.tif",
  EPSG = 4326,
  rios = "rios_ibge.shp",
  massa_dagua = "massa_dagua.shp",
  threshold = 250,
  min_length = 0.02,
  max_ordem = 4
)
}
\arguments{
\item{dir_shp}{Diretório dos dados espaciais}

\item{dir_dem}{Diretório da imagem tif srtm}

\item{dir_out}{Diretório de saída}

\item{limite}{nome do arquivo da área do projeto}

\item{area_urbana}{nome do arquivo da área urbana}

\item{estacoes}{Estações geradas a partir da drenagem modelada}

\item{srtm}{DEM da área}

\item{EPSG}{Sistema de coordenadas}

\item{rios}{nome do arquivo os rios}

\item{massa_dagua}{nome do arquivo da massa d'água}

\item{threshold}{Valor de corte da área das bacias modeladas}

\item{min_length}{Comprimento mínimo do curso do rio}

\item{max_ordem}{Ordem Strhaler máxima para a validaçãon da estação}
}
\value{
Imagem (.png) do mapa de planejamento preliminar.
Arquivo da shape (shp) das estações.
}
\description{
Desenvolvido para automatizar o processo de escolha do locais de amostragem
que devem ser a 100 m acima dos pontos de confluência dos rios. Também são
excluídos locais nas áreas urbanas e de massa d'água. Todos os arquivos devem
ser no formato shapefile.
}
\examples{
#gera_estacoes()

}
