% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersecta_bacias.R
\name{intersecta_bacias}
\alias{intersecta_bacias}
\title{Intersecta polígonos das bacias com outra feição}
\usage{
intersecta_bacias(
  limite = "carta_100M",
  bacias = "bacias_area",
  feicao = "geologia",
  legenda = "legenda",
  estacoes = "estacoes"
)
}
\arguments{
\item{limite}{Limite da área do projeto}

\item{bacias}{bacias obtidas a partir da função gera_bacias}

\item{feicao}{polígonos de algum tema (geologia,
solo, geofísica, geomorfologia, etc)}

\item{legenda}{Rótulos das unidades da feição}

\item{estacoes}{Estações de coleta}
}
\value{
Bacias
}
\description{
Cada bacia é recortada pelos polígonos da feição de sobreposição, que pode ser
de mapeamento.
}
\examples{
#intersecta_bacias()
}
