% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersecta_bacias.R
\name{intersecta_bacias}
\alias{intersecta_bacias}
\title{Intersecta polígonos das bacias com outra feição}
\usage{
intersecta_bacias(
  dir_in = "inputs/campo/",
  dir_out = "outputs/",
  limite = "carta_100M",
  bacias = "bacias_area",
  feicao = "geologia",
  tipo_leg = 1,
  estacoes = "estacoes"
)
}
\arguments{
\item{dir_in}{Diretório dos dados espaciais e legenda}

\item{dir_out}{Diretório de saída onde serão gravados as planilhas .csv}

\item{limite}{Limite da área do projeto}

\item{bacias}{bacias obtidas a partir da função gera_bacias}

\item{feicao}{polígonos de algum tema (geologia,
solo, geofísica, geomorfologia, etc)}

\item{tipo_leg}{Tipo de legenda 1= pela SIGLA 2 = pelo RANGE}

\item{estacoes}{Estações de coleta}
}
\value{
Bacias
}
\description{
Cada bacia é recortada pelos polígonos da geologia. Esta função está
configurada para a shape do mapa geológico ao milhonésimo 2014 do SGB-CPRM.
}
\examples{
#intersecta_bacias()
}
