% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersecta_bacias.R
\name{intersecta_bacias}
\alias{intersecta_bacias}
\title{Intersecta polígonos das bacias com outra feição}
\usage{
intersecta_bacias(
  dir_in = "inputs/campo/",
  dir_out = "outputs/",
  bases_model,
  bacias,
  feicao_rec,
  tipo_leg = 2,
  estacoes,
  classe_am = 2,
  nome_xml = "geologia"
)
}
\arguments{
\item{dir_out}{Diretório de saída onde serão gravados as planilhas .csv}

\item{bacias}{bacias obtidas a partir da função gera_bacias}

\item{feicao_rec}{polígonos de algum tema (geologia,
solo, geofísica, geomorfologia, etc)}

\item{tipo_leg}{Tipo de legenda 1= pela SIGLA 2 = pelo RANGE}

\item{estacoes}{Estações de coleta}

\item{classe_am}{Classe da amostra: 1 = concentrado de bateia, 2 = sedimento de
corrente, 3 = rocha, 4 = solo, 5 = água}

\item{nome_xml}{}

\item{limite}{Limite da área do projeto}
}
\value{
Bacias
}
\description{
Cada bacia é recortada pelos polígonos da geologia. Esta função está
configurada para a shape do mapa geológico ao milhonésimo 2014 do SGB-CPRM.
}
\examples{
#intersecta_bacias()
}
