% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/le_boletim_mineral.R
\name{le_boletim_mineral}
\alias{le_boletim_mineral}
\title{Lê boletins de análises mineralométricas}
\usage{
le_boletim_mineral(dir_bol)
}
\arguments{
\item{dir_bol}{Diretório dos boletins analíticos ex: "inputs/mineral/B/"}
}
\value{
Retorna uma lista com todos os dados do boletim: resultados
analíticos, condições analíticas, requisição das análises, preparação das
amostras, separação grav-magnética.
}
\description{
Lê boletins de análises mineralométricas
}
\examples{
#le_boletim_mineral()
}
