% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/le_boletim_mineral.R
\name{le_boletim_mineral}
\alias{le_boletim_mineral}
\title{Lê boletins de análises mineralométricas}
\usage{
le_boletim_mineral(tipo = 1)
}
\arguments{
\item{tipo}{Classe da amostra: 1 = concentrado de bateia, 2 = sedimento de
corrente, 3 = rocha, 4 = solo}
}
\value{
Retorna uma lista com todos os dados do boletim: resultados
analíticos, condições analíticas, requisição das análises, preparação das
amostras, separação grav-magnética.
}
\description{
Lê boletins de análises mineralométricas
}
\examples{
#le_boletim_mineral()
}
