% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/le_boletim_quimica.R
\name{le_boletim_quimica}
\alias{le_boletim_quimica}
\title{Lê boletins de análises químicas}
\usage{
le_boletim_quimica(tipo = 2)
}
\arguments{
\item{tipo}{Classe da amostra: 1 = concentrado de bateia, 2 = sedimento de
corrente, 3 = rocha, 4 = solo, 5 = água}
}
\value{
Retorna uma lista com todos os dados do boletim: resultados
analíticos, condições analíticas, QA/QC, requisição das análises e
preparação das amostras.
}
\description{
Esta função lê os boletins csv e extrai as informações na forma de tabelas
que são gravadas no diretório de saída.
}
\examples{
# le_boletim_quimica()
}
