% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapa_tif_unidades.R
\name{mapa_tif_unidades}
\alias{mapa_tif_unidades}
\title{Mapa geoquímico (TIF)}
\usage{
mapa_tif_unidades(
  bacias,
  geologia,
  dados_transf,
  info_bol,
  litho_max,
  nbc = 10,
  elem_val,
  tipo_leg,
  leg
)
}
\arguments{
\item{leg}{}
}
\description{
Mapa geoquímico (TIF)
}
