% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modela_bacias.R
\name{modela_bacias}
\alias{modela_bacias}
\title{Extrai a drenagem do SRTM}
\usage{
modela_bacias(
  fase = 2,
  EPSG = 4326,
  dem,
  bases_model,
  terreno,
  ex_campo,
  classe_am,
  fonte_shp = FALSE,
  arquivo_shp,
  bacia_minima = 4,
  bacia_maxima = 100,
  tipo = 1,
  funcao_snap = 1,
  snap_dist = "0.02",
  decrescente = FALSE,
  max_ordem = 3,
  dir_out = "outputs/",
  wbt_wd = "outputs/modelo/"
)
}
\arguments{
\item{fase}{Fase do levantamento geoquímico (1= Planejamento, 2 = Pós campo)}

\item{EPSG}{crs da camada}

\item{dem}{Modelo digital de terreno (ex. SRTM)}

\item{bases_model}{Bases para a modelagem das bacias e drenagem pela
função gera_bases_model}

\item{terreno}{Objeto da saída da função modela_terreno}

\item{ex_campo}{Dados de campo extraídos pela função extrai_dados_campo}

\item{classe_am}{Classe da amostra: 1 = concentrado de bateia, 2 = sedimento de
corrente}

\item{fonte_shp}{Valor lógico se oas esstações forem shapefile}

\item{arquivo_shp}{Caso for shape fornecer o caminho do arquivo}

\item{bacia_minima}{Área mínima das bacias planejadas}

\item{bacia_maxima}{Área máxima das bacias planejadas}

\item{tipo}{Tipo de processamento do deslocamento 1: Deslocamento sem controle,
2: controlado pela ordem da drenagem e 3: sem deslocamento.
Para o tipo 1 informar o snap_disp.
Para o tipo 2 informar o snap_dist e o valor lógico decrescente.}

\item{funcao_snap}{1: wbt_jenson_snap_pour_points 2: wbt_snap_pour_points}

\item{snap_dist}{Deslocamento máximo do ponto até a drenagem}

\item{decrescente}{Logico TRUE para ordem decrecente e FALSE para ordem crescente.}

\item{max_ordem}{Máxima ordem do rio para a busca do snap point}

\item{dir_out}{Diretório de saída}

\item{wbt_wd}{diretório onde estão salvos os dados da modelagem do DEM}
}
\value{
Rasters da preparação das imagens para a extração do modelo de
drenagem. Shapes e rasters das drenagens e trechos drenagems com as ordens
de strahler.
}
\description{
A partir da imagem SRTM, rios, massa d'água é gerado um modelo de drenagem.
A densidade de drenagem é dada pelo valor de threshold, quanto maior
este valor, menor a densidade dos cursos dos rios.
}
\examples{
# modelo_bacias()
}
