% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modela_bacias.R
\name{modela_bacias}
\alias{modela_bacias}
\title{Extrai a drenagem do SRTM}
\usage{
modela_bacias(
  fase = 2,
  EPSG = 4326,
  bacia_minima = 4,
  bacia_maxima = 100,
  srtm = "inputs/imagens/srtm.tif",
  rios = "inputs/campo/rios_ibge.shp",
  massa_dagua = "inputs/campo/massa_dagua.shp",
  limite = "inputs/campo/carta_100M.shp",
  estacoes = "inputs/campo/estacoes.shp",
  threshold = 250,
  snap_dist = 0.02,
  min_length = 0.02,
  max_ordem = 3
)
}
\arguments{
\item{fase}{Fase do levantamento geoquímico (1= Planejamento, 2 = Pós campo)}

\item{EPSG}{crs da camada}

\item{bacia_minima}{Área mínima das bacias planejadas}

\item{bacia_maxima}{Área máxima das bacias planejadas}

\item{srtm}{Imagem de radar com os dados de altimetria}

\item{rios}{Shape dos rios obtidos por cartas digitais, fotografia aérea ou
dados do IBGE}

\item{massa_dagua}{Shape da massa d'água, obtidos por cartas digitais,
fotografia aérea ou dados do IBGE}

\item{limite}{Polígonono da área do projeto}

\item{estacoes}{Estaçãoes de coleta}

\item{threshold}{Numero de células mínimo para o traçado das drenagens}

\item{snap_dist}{Deslocamento máximo do ponto até a drenagem}

\item{min_length}{Comprimento mínimo da drenagem}

\item{max_ordem}{Máxima ordem do rio para a busca do snap point}
}
\value{
Rasters da preparação das imagens para a extração do modelo de
drenagem. Shapes e rasters das drenagens e trechos drenagems com as ordens
de strahler.
}
\description{
A partir da imagem SRTM, rios, massa d'água é gerado um modelo de drenagem.
A densidade de drenagem é dada pelo valor de threshold, quanto maior
este valor, menor a densidade dos cursos dos rios.
}
\examples{
# modelo_bacias()
}
