% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepara_bases.R
\name{prepara_bases}
\alias{prepara_bases}
\title{Prepara a base de dados para o processamento dos dados geoquímicos}
\usage{
prepara_bases(analise, tipo, fcampo = "fcampo")
}
\arguments{
\item{analise}{Selecione o tipo de análise: 1 = mineral ou 2 = química}

\item{tipo}{Selecione a classe da amostra: 1 = concentrado de bateia, 2 =
"sedimento de corrente", 3 = solo, 4 = rocha, 5 = água}

\item{fcampo}{Nome do arquivo acess mdb com dados de campo.}
}
\value{
Retorna a lista com as bases de dados em diferentes formatos
transformados ou brutos, pivotados ou não.
}
\description{
Recupera e une os dados dos boletins analíticos e os dados de campo da base de
dados fcampo. Os dados precisam estar nos diretórios gerados pela função
estrutura_diretorio. Os boletins analíticos devem estar no formato .csv com
layout modelo nos diretórios inputs/mineral ou inputs/quimica, conforme o
exemplo.
}
\examples{
#prepara_bases(analise = 2, tipo = 2)
}
