% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepara_bases.R
\name{prepara_bases}
\alias{prepara_bases}
\title{Prepara a base de dados para o processamento dos dados geoquímicos}
\usage{
prepara_bases(dir_bol, classe_am, analise, quimica, mineral, ex_campo)
}
\arguments{
\item{dir_bol}{Diretório dos boletins analíticos ex: "inputs/qumica/R/"}

\item{classe_am}{Classe da amostra: 1 = concentrado de bateia, 2 = sedimento de
corrente, 3 = rocha, 4 = solo, 5 = água#'}

\item{analise}{Selecione o tipo de análise: 1 = mineral ou 2 = química}
}
\value{
Retorna a lista com as bases de dados em diferentes formatos
transformados ou brutos, pivotados ou não.
}
\description{
Recupera e une os dados dos boletins analíticos e os dados de campo da base de
dados fcampo. Os dados precisam estar nos diretórios gerados pela função
estrutura_diretorio. Os boletins analíticos devem estar no formato .csv com
layout modelo nos diretórios inputs/mineral ou inputs/quimica, conforme o
exemplo.
}
