% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepara_bases.R
\name{prepara_bases}
\alias{prepara_bases}
\title{Prepara a base de dados para o processamento dos dados geoquímicos}
\usage{
prepara_bases(
  dir_bol,
  classe_am,
  analise,
  dir_base,
  tipo_base,
  base_campo,
  dir_out,
  ref_ucc,
  EPSG
)
}
\arguments{
\item{dir_bol}{Diretório dos boletins analíticos ex: "inputs/qumica/R/"}

\item{classe_am}{Classe da amostra: 1 = concentrado de bateia, 2 = sedimento de
corrente, 3 = rocha, 4 = solo, 5 = água}

\item{analise}{Selecione o tipo de análise: 1 = mineral ou 2 = química}

\item{dir_base}{Diretório dos dados de campo}

\item{tipo_base}{tipo de base de dados: 1 = FCAMPO, 2 = SURVEY123, 3 = QFIELD}

\item{base_campo}{Nome do arquivo das bases de campo}

\item{dir_out}{Diretório de saída para os dados de campo}

\item{ref_ucc}{Referência Concentração Média da Crosta Superior
Rudnick e Gao, 2004}

\item{EPSG}{Sistema de coordenadas}
}
\value{
Retorna a lista com as bases de dados em diferentes formatos
transformados ou brutos, pivotados ou não.
}
\description{
Recupera e une os dados dos boletins analíticos e os dados de campo da base de
dados fcampo. Os dados precisam estar nos diretórios gerados pela função
estrutura_diretorio. Os boletins analíticos devem estar no formato .csv com
layout modelo nos diretórios inputs/mineral ou inputs/quimica, conforme o
exemplo.
}
\examples{
#prepara_bases(dir_bol = "inputs/quimica/S/",classe_am = 2,analise = 2,
dir_base = "inputs/campo/", tipo_base = 1, base_campo = "fcampo" )
}
