% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepara_dem.R
\name{prepara_dem}
\alias{prepara_dem}
\title{Title}
\usage{
prepara_dem(
  limite = "inputs/campo/area_srtm.shp",
  EPSG = 4326,
  z = 11,
  cellsize = 0.000277777777674848
)
}
\arguments{
\item{limite}{Limite da área do srtm}

\item{EPSG}{Projeçao usada}

\item{z}{Nível de zoom para recuperar o SRTM}

\item{cellsize}{}
}
\description{
Title
}
