% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepara_dem.R
\name{prepara_dem}
\alias{prepara_dem}
\title{Prepara imagem de SRTM para a modelagem de terreno}
\usage{
prepara_dem(
  limite = "inputs/campo/area_srtm.shp",
  dir_out = "inputs/imagens/",
  EPSG = 4326,
  z = 11
)
}
\arguments{
\item{limite}{Limite da área do srtm}

\item{dir_out}{Diretório de saída}

\item{EPSG}{Projeçao usada}

\item{z}{Nível de zoom para recuperar o SRTM}
}
\description{
Recupera dados de elevação do terreno da Amazon Web Services Terrain Tile
didoníveis em https://registry.opendata.aws/terrain-tiles.
Para mais detalhes visite o repositório https://github.com/jhollist/elevatr
}
