% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepara_dem.R
\name{prepara_dem}
\alias{prepara_dem}
\title{Prepara imagem de SRTM para a modelagem de terreno}
\usage{
prepara_dem(
  dir_in = "inputs/campo/",
  limite_srtm = "area_srtm.shp",
  dir_out = "outputs/",
  EPSG = 4326,
  z = 11
)
}
\arguments{
\item{limite_srtm}{}

\item{dir_out}{Diretório de saída}

\item{EPSG}{Projeçao usada}

\item{z}{Nível de zoom para recuperar o SRTM}
}
\description{
Recupera dados de elevação do terreno da Amazon Web Services Terrain Tile
didoníveis em https://registry.opendata.aws/terrain-tiles.
Para mais detalhes visite o repositório https://github.com/jhollist/elevatr
}
