% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepara_legenda.R
\name{prepara_legenda}
\alias{prepara_legenda}
\title{Cria legenda apartir da shape da feição}
\usage{
prepara_legenda(
  file_shp = "inputs/campo/geologia.shp",
  file_xml = "inputs/diversos/geologia.xml"
)
}
\arguments{
\item{file_shp}{Endereço e nome do arquivo da geologia}

\item{file_xlm}{Endereço e nome do arquivo xlm}
}
\value{
Tabelas com atrubutos da shape da geologia e cores das unidades
(col_hex) e índice da geologia (Geo_Reg) em duas opções uma baseada no campo
SIGLA e outra baseada no campo RANGE.
}
\description{
Este processamento usa a shape e xml do mapa das unidades litoestratigráficas
do Brasil (2014) obtidos do GEOSGB (SGB-CPRM).
}
\examples{
#legenda <- prepara_legenda()
#legenda
}
