% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepara_legenda.R
\name{prepara_legenda}
\alias{prepara_legenda}
\title{Cria legenda apartir da shape da feição}
\usage{
prepara_legenda(
  file_shp = "inputs/campo/geologia.shp",
  file_xml = "inputs/diversos/geologia.xml"
)
}
\arguments{
\item{file_shp}{Endereço e nome do arquivo da geologia}

\item{file_xlm}{Endereço e nome do arquivo xlm}
}
\value{
Tabela com atrubutos da shape da geologia e cores das unidades (col_hex)
e índice da geologia (Geo_Reg)
}
\description{
Cria legenda apartir da shape da feição
}
\examples{
#legenda <- prepara_legenda()
#legenda
}
