% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaqc_geoq.R
\name{qaqc_gq}
\alias{qaqc_gq}
\title{Controle de qualidade analitico e amostral}
\usage{
qaqc_gq(dir_out, base)
}
\arguments{
\item{dir_out}{Diretório de saída}

\item{base}{base de dados analíticos}
}
\description{
Controle de qualidade analitico e amostral
}
