% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resgata_legenda_xml.R
\name{resgata_legenda_xml}
\alias{resgata_legenda_xml}
\title{Extrai dados da legenda de um arquivo de estilo xml}
\usage{
resgata_legenda_xml(dir_xml = "inputs/campo/", nome_xml = "geologia")
}
\arguments{
\item{dir_xml}{Diretório do arquivo xml}

\item{nome_xml}{Nome do arquivo xml}
}
\description{
No QGIS use a ferramenta Create style database from project para gerar o XML,
da feição pretendida, aqui usei usei o do mapa geológico do Brasil
ao milhonésimo de 2014 do SGB-CPRM.
}
