% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resgata_legenda_xml.R
\name{resgata_legenda_xml}
\alias{resgata_legenda_xml}
\title{Extrai dados da legenda de um arquivo de estilo xml}
\usage{
resgata_legenda_xml(
  file_xml,
  find1 = "//symbol/layer",
  indice = "id",
  value_cor = "value",
  find2 = "//symbol",
  value_name = "name",
  value_sub = "geologia "
)
}
\arguments{
\item{find1}{Nó de busca para a cor}

\item{indice}{Índice dos registros da legenda xml}

\item{value_cor}{Nome da variável com as cores}

\item{find2}{Nó de busca para os nomes das siglas}

\item{value_name}{Nome da variável dos nomes das siglas}

\item{value_sub}{Caracteres para substituir no nome das siglas}

\item{file}{Endereço e nome do arquivo}
}
\description{
Extrai dados da legenda de um arquivo de estilo xml
}
