% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_estatistica.R
\name{tabela_estatistica}
\alias{tabela_estatistica}
\title{Sumário Estatístico das análises químicas}
\usage{
tabela_estatistica(
  tipo = 1,
  rotulo_lito = "outputs/mylitho.csv",
  estacao = "outputs/estacoes.shp",
  ref_ucc = "inputs/quimica/ucc.csv",
  dir_bol = "inputs/quimica/S/",
  classe_am = 2,
  analise = 2,
  dir_base = "inputs/campo/",
  tipo_base = 1,
  base_campo = "fcampo",
  dir_out = "outputs/"
)
}
\arguments{
\item{tipo}{Modo de processamento: 1 = Todas amaostras e
2 = agrupadas por unidades}

\item{rotulo_lito}{Planilha gerada pela função intersecta_bacias com as
unidades litológicas dominantes em cada bacia.}

\item{estacao}{Dados de campo das estações amostradas (shp)}

\item{ref_ucc}{Planilha com valores da concentração média da crosta superior
Rudnick e Gao 2004}

\item{dir_bol}{Diretório dos arquivos dos boletins}

\item{classe_am}{Classe de amostra}

\item{analise}{tipo de análise 1 = mineralógica e 2 = química}

\item{dir_base}{Diretório da base de dados de campo}

\item{tipo_base}{Tipo de base de campo 1 = FCAMP, 2 = SURVEY123 e 3 = QFIELD}

\item{base_campo}{Nome do arquivo da base de campo (sem extenção)}

\item{dir_out}{Diretório de saída dados de campo e estações}
}
\description{
Gera planilhas cos os principais parâmetros estatísticos
}
