% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testes_statisticos.R
\name{testes_estatisticos}
\alias{testes_estatisticos}
\title{Testes Estatíticos}
\usage{
testes_estatisticos(mydata, mydata_b, info_bol, mylitho_max_sc)
}
\arguments{
\item{mydata}{Dados químicos transformados}

\item{mydata_b}{Dados quimícos brutos}

\item{info_bol}{Incormações das condições analíticas}

\item{mylitho_max_sc}{}
}
\description{
Testes de normalidade e de comparação das amostras
}
